updating

+++ app/Http/Controllers/CommentController.php
@@ -6,12 +6,14 @@ use App\Enums\ErrorCodeEnum;
 use App\Http\Requests\Comment\StoreCommentRequest;
 use App\Http\Requests\Comment\UpdateCommentRequest;
 use App\Http\Resources\CommentResource;
+use App\Models\Comment;
 use App\Repositories\CommentRepository;
 use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\HttpException;
 
 class CommentController extends Controller
 {
@@ -48,6 +50,9 @@ class CommentController extends Controller
     {
         DB::beginTransaction();
         try {
+            if ($request->user()->cannot('store', Comment::class)) {
+                throw new HttpException(Response::HTTP_FORBIDDEN);
+            }
             $requestData = $request->validated();
             $comment = $this->commentRepository->create($requestData);
             $commentResource = new CommentResource($comment);
@@ -62,8 +67,17 @@ class CommentController extends Controller
     public function getByClubSession(Request $request, string $id): JsonResponse
     {
         $conditions = $request->all();
-        $comments = $this->commentRepository->getCommentList([...$conditions,'session_code' => $id]);
+        $comments = $this->commentRepository->getCommentList([...$conditions, 'session_code' => $id]);
         return $this->sendPaginationResponse($comments, CommentResource::collection($comments));
+    }
+
+    public function show(string $id): JsonResponse
+    {
+        $comment = $this->commentRepository->getComment($id);
+        if(!$comment) {
+            return $this->sendError(__('common.not_found'), ErrorCodeEnum::CommentShow, Response::HTTP_NOT_FOUND);
+        }
+        return $this->sendResponse($comment, __('common.get_success'));
     }
 
     /**
@@ -78,11 +92,13 @@ class CommentController extends Controller
         DB::beginTransaction();
         try {
             $requestData = $request->validated();
-            $comment = $this->commentRepository->find($id);
+            $comment = $this->commentRepository->getComment($id);
             if (!$comment) {
                 return $this->sendError(__('common.not_found'), ErrorCodeEnum::CommentUpdate, Response::HTTP_NOT_FOUND);
             }
-
+            if ($request->user()->cannot('update', $comment)) {
+                throw new HttpException(Response::HTTP_FORBIDDEN);
+            }
             $comment = $this->commentRepository->update($id, $requestData);
             $commentResource = new CommentResource($comment);
             DB::commit();
@@ -99,13 +115,16 @@ class CommentController extends Controller
      * @param string $id
      * @return JsonResponse
      */
-    public function destroy(string $id): JsonResponse
+    public function destroy(Request $request, string $id): JsonResponse
     {
         DB::beginTransaction();
         try {
             $comment = $this->commentRepository->find($id);
             if (!$comment) {
                 return $this->sendError(__('common.not_found'), ErrorCodeEnum::CommentDelete, Response::HTTP_NOT_FOUND);
+            }
+            if ($request->user()->cannot('destroy', $comment)) {
+                throw new HttpException(Response::HTTP_FORBIDDEN);
             }
             $this->commentRepository->delete($id);
             DB::commit();
